requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pandas==2.2.3
polars==1.17.1
pyarrow==18.1.0
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.3
plotly==5.24.1
pyyaml==6.0.2
python-multipart==0.0.18
joblib==1.4.2
